action-traction/pyproject.toml
[tool.poetry]
name = "action-traction"
version = "0.1.0"
description = ""
authors = ["MaddyKapfhammer <maddy.kapfhammer@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
action-traction = "action_traction.main:app"

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.3.2"
GitPython = "^3.1.17"
PyDriller = "^2.0"
pandas = "^1.2.4"
numpy = "^1.20.3"
statistics = "^1.0.3"
PyYAML = "^5.4.1"
nested-lookup = "^0.2.22"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
